+++ lua/render-markdown/config/bullet.lua
@@ -48,7 +48,7 @@ M.default = {
     ordered_icons = function(ctx)
         local value = vim.trim(ctx.value)
         local index = tonumber(value:sub(1, #value - 1))
-        return string.format('%d.', index > 1 and index or ctx.index)
+        return ('%d.'):format(index > 1 and index or ctx.index)
     end,
     -- Padding to add to the left of bullet point.
     -- Output is evaluated depending on the type.